Todoist: Bring back various commands and bug fixes
