--- extensions/todoist/src/components/LabelTasks.tsx
@@ -1,73 +1,0 @@
-import { List, ActionPanel, Action, Icon } from "@raycast/api";
-import { useMemo } from "react";
-
-import CreateTask from "../create-task";
-import { labelSort } from "../helpers/labels";
-import { QuickLinkView, ViewMode } from "../home";
-import useCachedData from "../hooks/useCachedData";
-import useViewTasks from "../hooks/useViewTasks";
-
-import CreateViewAction from "./CreateViewAction";
-import TaskListItem from "./TaskListItem";
-import TaskListSections from "./TaskListSections";
-
-type LabelTasksProps = { name: string; quickLinkView?: QuickLinkView };
-
-function LabelTasks({ name, quickLinkView }: LabelTasksProps) {
-  const [data] = useCachedData();
-
-  const tasks = useMemo(() => {
-    const labelTasks = data?.items.filter((task) => task.labels.includes(name)) ?? [];
-    return labelSort(labelTasks);
-  }, [data, name]);
-
-  const { sections, viewProps, sortedTasks } = useViewTasks(`todoist.label${name}`, { tasks, data });
-
-  if (tasks.length === 0) {
-    return (
-      <List.EmptyView
-        title="No tasks for this label."
-        description="How about adding one?"
-        actions={
-          <ActionPanel>
-            <Action.Push
-              title="Create Task"
-              icon={Icon.Plus}
-              target={<CreateTask fromLabel={name} />}
-              shortcut={{ modifiers: ["cmd", "shift"], key: "a" }}
-            />
-
-            {quickLinkView ? <CreateViewAction {...quickLinkView} /> : null}
-          </ActionPanel>
-        }
-      />
-    );
-  }
-
-  return (
-    <>
-      {viewProps.groupBy?.value === "default" ? (
-        sortedTasks.map((task) => {
-          return (
-            <TaskListItem
-              key={task.id}
-              task={task}
-              mode={ViewMode.search}
-              viewProps={viewProps}
-              quickLinkView={quickLinkView}
-            />
-          );
-        })
-      ) : (
-        <TaskListSections
-          mode={ViewMode.project}
-          sections={sections}
-          viewProps={viewProps}
-          quickLinkView={quickLinkView}
-        />
-      )}
-    </>
-  );
-}
-
-export default LabelTasks;
