+++ extensions/todoist/src/components/TaskListItem.tsx
@@ -2,14 +2,13 @@ import { ActionPanel, Icon, List, Action, Color } from "@raycast/api";
 import { format } from "date-fns";
 import removeMarkdown from "remove-markdown";
 
-import { Task } from "../api";
+import { SyncData, Task } from "../api";
 import { getCollaboratorIcon } from "../helpers/collaborators";
 import { getColorByKey } from "../helpers/colors";
 import { isRecurring, displayDueDate, isExactTimeTask, displayDueDateTime, isOverdue } from "../helpers/dates";
 import { getPriorityIcon, priorities } from "../helpers/priorities";
 import { displayReminderName } from "../helpers/reminders";
-import { QuickLinkView, ViewMode } from "../home";
-import useCachedData from "../hooks/useCachedData";
+import { ViewMode } from "../helpers/tasks";
 import { ViewProps } from "../hooks/useViewTasks";
 
 import TaskActions from "./TaskActions";
@@ -19,12 +18,11 @@ type TaskListItemProps = {
   task: Task;
   mode?: ViewMode;
   viewProps?: ViewProps;
-  quickLinkView?: QuickLinkView;
+  data?: SyncData;
+  setData: React.Dispatch<React.SetStateAction<SyncData | undefined>>;
 };
 
-export default function TaskListItem({ task, mode, viewProps, quickLinkView }: TaskListItemProps): JSX.Element {
-  const [data] = useCachedData();
-
+export default function TaskListItem({ task, mode, viewProps, data, setData }: TaskListItemProps): JSX.Element {
   const taskComments = data?.notes.filter((note) => note.item_id === task.id);
   const accessories: List.Item.Accessory[] = [];
   const keywords: string[] = [];
@@ -57,7 +55,7 @@ export default function TaskListItem({ task, mode, viewProps, quickLinkView }: T
 
   if (task.labels && task.labels.length > 0) {
     accessories.unshift({
-      icon: { source: "tag.svg", tintColor: Color.SecondaryText },
+      icon: { source: Icon.Tag, tintColor: Color.SecondaryText },
       text: `${task.labels.length}`,
       tooltip: `${task.labels.join(", ")}`,
     });
@@ -146,7 +144,7 @@ export default function TaskListItem({ task, mode, viewProps, quickLinkView }: T
         <ActionPanel title={task.content}>
           <Action.Push title="Show Details" target={<TaskDetail taskId={task.id} />} icon={Icon.Sidebar} />
 
-          <TaskActions task={task} viewProps={viewProps} mode={mode} quickLinkView={quickLinkView} />
+          <TaskActions task={task} viewProps={viewProps} mode={mode} data={data} setData={setData} />
         </ActionPanel>
       }
     />
